refactor(http): extraer factory.rs con las factorías de cliente (#444)
Mueve build_wreq_client, create_http_client, create_http_client_with_config y
get_random_user_agent_from_pool (más las constantes CLIENT_HINTS) a factory.rs.
client.rs conserva un pub use de las factorías para que mod.rs y los tests sigan
resolviendo sin cambios; build_wreq_client queda pub(super) para HttpClient::new.

API pública preservada vía re-exports; cero churn en call sites (crawler, sitemap,
url_validator, preflight, container). Workspace completo compila.
