+++ crates/webfang_core/src/application/http_client/mod.rs
@@ -37,6 +37,7 @@
 //! ```
 
 mod client;
+mod factory;
 pub mod port;
 mod retry;
